rename file for consistency with class name
